EV_Stations/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
class-registry==2.1.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-ratelimit==4.1.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework-oauth==1.1.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filters==1.3.2
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.5.4
mysql-connector-python==9.3.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
timedelta==2020.12.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
gunicorn
psycopg2-binary
python-decouple
